# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ToDo API",
    "description": "An ASP.NET Core Web API for managing ToDo items",
    "termsOfService": "https://example.com/terms",
    "contact": {
      "name": "Example Contact",
      "url": "https://example.com/contact"
    },
    "license": {
      "name": "Example License",
      "url": "https://example.com/license"
    },
    "version": "v1"
  },
  "paths": {
    "/api/Attach": {
      "get": {
        "tags": [
          "Attach"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Meta"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Meta"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Meta"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Attach"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Meta"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Meta"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Meta"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Meta"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meta"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meta"
                }
              }
            }
          }
        }
      }
    },
    "/api/Attach/{id}": {
      "get": {
        "tags": [
          "Attach"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Meta"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meta"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meta"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Attach"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Meta"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Meta"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Meta"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Attach"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Dev/RecreateDatabase": {
      "post": {
        "tags": [
          "Dev"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Dev/EnsureTempFilesFolder": {
      "post": {
        "tags": [
          "Dev"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Files/GetFile": {
      "get": {
        "tags": [
          "Files"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Files/Upload": {
      "post": {
        "tags": [
          "Files"
        ],
        "parameters": [
          {
            "name": "Content-Range",
            "in": "header",
            "required": true,
            "example": "123"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Files/UploadBig": {
      "post": {
        "tags": [
          "Files"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            },
            "text/json": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Files/{id}": {
      "put": {
        "tags": [
          "Files"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Files"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/SomeTasks": {
      "get": {
        "tags": [
          "SomeTasks"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SomeTask"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SomeTask"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SomeTask"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "SomeTasks"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SomeTask"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SomeTask"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SomeTask"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SomeTask"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SomeTask"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SomeTask"
                }
              }
            }
          }
        }
      }
    },
    "/api/SomeTasks/{id}": {
      "get": {
        "tags": [
          "SomeTasks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SomeTask"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SomeTask"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SomeTask"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "SomeTasks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SomeTask"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SomeTask"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SomeTask"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "SomeTasks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EStatus": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
      "Meta": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "fileDir": {
            "type": "string",
            "nullable": true
          },
          "fileNameOriginal": {
            "type": "string",
            "nullable": true
          },
          "resourceId": {
            "type": "string",
            "format": "uuid"
          },
          "fileExt": {
            "type": "string",
            "nullable": true
          },
          "someTaskId": {
            "type": "integer",
            "format": "int32"
          },
          "isUploaded": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "SomeTask": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "dateTimeCreation": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/EStatus"
          },
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Meta"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Please enter a valid token",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}